Add separate range messages for “now” endpoints
Instead of interpolating the current date into the regular “range”
violation messages, introduce separate messages for all cases where one
range endpoint can be “now”. This should hopefully make the messages
easier to understand and also makes the messages themselves independent
from the point in time in which they’re generated (though their validity
still varies by date, of course).

To allow RangeChecker to distinguish between a “now” parameter and one
that just happens to match the current date, introduce a NowValue
subclass of TimeValue and make ConstraintParameterParser return that
instead of a TimeValue for the current time.

Bug: T187058
Change-Id: If0ce863ef49b117ed20b3b9c9d07de2ceb0bc353