Add future time to DependencyMetadata
A value can now not only depend on a set of entity IDs, but also on the
fact that a certain point in time is in the future, not in the past.

This doesn’t add a corresponding method for depending on a certain point
in time being in the *past*, because the only way that this could cause
a value to be invalid is if the system clock jumps backwards (and likely
by a tremendous amount, given that we expect to use this kind of
dependency with “day” or even “month” granularity), and in that case
it’s not clear that the new time is more correct than the old time, and
that it would therefore be correct to invalidate the value.

Bug: T187061
Change-Id: I37e131a07b0574cc4225cae03baf0c8d20aaf609