Add ConstraintParameterException::getViolationMessage()
Currently, it just delegates to getMessage(), but soon a
ConstraintParameterException can also contain a ViolationMessage (and
support for plain strings will eventually be removed). By declaring this
method up-front we can already prepare some places that use a
ConstraintParameterException for this possibility without risk of
breakage (because it’s not yet possible to instantiate a
ConstraintParameterException where getMessage() will actually return a
ViolationMessage).

Change-Id: Id7eca131b132e59f44a850fa2178a8ad5980ce9a