Remove plain message CPE serializations from cache map
Any time we refresh the cache map for a given regex, remove all entries
that are ConstraintParameterException serializations with a plain
message instead of a serialized ViolationMessage. This ensures that we
can at some point remove support for plain strings in
ConstraintParameterException without being tripped up by old
serializations in the cache.

The proper hashes in the test array are SHA2-256 digests of the strings
“Wikidata = ♥”, “Wikidata = <3”, and “non-bool value” (without quotation
marks and without trailing newlines).

Bug: T187186
Change-Id: I49d41708140fc7a984c81183c91711eb11925985