Remove most of ConstraintParameterRenderer
Now that all violation messages are rendered by
ViolationMessageRenderer, most of ConstraintParameterRenderer’s methods
can be removed – only the parts that are necessary for the rendering of
check result parameters in the special page remain. This includes the
removal of role parameters even from methods that are still necessary.

formatItemIdSnakValue is also simplified to remove the HTML wrapping,
since it is now only used in the context of
SpecialConstraintReport::buildExpandableElement, which HTML-escapes the
content. (Previously, the whole “<span …>”-wrapping was visible as plain
text on the special page, which I hadn’t noticed.)

Bug: T187201
Change-Id: I154d9cdebe032054afdb8e6191c7dca930ba7fec