Correctly handle errors from metadata accessor
WikiPageEntityMetaDataLookup::loadRevisionInformation can return false
elements in the array, even though its interface (…Accessor) does not
specify this. Make sure that we don’t store anything into the cache in
that case.

The interface change to CachingResultsBuilder::getLatestRevisionIds
(allowing it to signal “skip caching”) will also be useful for T188312.

Change-Id: I1e5e27691ee97c182b5f67bbf3e2b6f24aeb0de8