Don’t send huge requests to metadata accessor
If there’s an attempt to load the revision information of an exorbitant
number of entities, just abort caching entirely. This should never
happen, but it’s intended as a safeguard to avoid database queries for
hundreds of thousands of pages, just in case the dependency information
gets messed up unexpectedly.

If this happens when called from getAndStoreResults(), we don’t store
the results, but still return them. If it happens in getStoredResults()
(even though such a value should never have ended up in the cache in the
first place), we behave as if one of the entities changed in the
meantime, and initiate a fresh constraint check.

Bug: T188312
Change-Id: I4187ceaf7f473576be4f4bea0241aac6a94d824c