Replace CheckResult’s Context with ContextCursor
A CheckResult now holds only a ContextCursor instead of a full Context.
It can still be instantiated with a full Context, in which case it will
get the cursor from it and also store the snak type and data value,
which are otherwise unavailable (the getters are changed to possibly
return null as well). The special case is updated to handle
possibly-null snak type and data value, although it should never happen
(this is only expected for CheckResults deserialized from cache, and the
special page doesn’t use caching).

Bug: T185712
Change-Id: I6b1d7533964144adea98ddc67da8c71bf896df29