Only log cache misses for nonempty entity ID lists
Even though the current implementation of logCheckConstraintsCacheMisses
supports being called with an empty list, it doesn’t really make sense
to say that there were cache misses for an empty list of entity IDs.
It’s cleaner to only call that method if the remaining set of entity IDs
really is nonempty.

Change-Id: I9707e733dafc67342ee557d3e1ad9b7c29ca2019