Add and use CheckResult::getContextCursor()
Add a getter for the cursor of a CheckResult’s context, and use that
instead of getContext() everywhere. This should let us remove the full
Context from CheckResult.

Some of the tests need larger restructurings for this, since they were
comparing directly against the Context and now need to compare against
something else.

Bug: T185712
Change-Id: Icb44b54fedd73c18f21100cdf700ee1d26b1a263