Reorganize Special:ConstraintReport
Replace the “claim” column (with the property and value) with two new
columns, one for the property ID and one for the violation message
(which is no longer a tooltip on the status). Results which don’t have a
message (compliance, not-in-scope, etc.) are no longer shown in the
table. (However, their quantity is still included in the summary above
the table.)

Bug: T188728
Change-Id: I3431154f13b42e355ce957465269d6d92c9e6277