Add more getters to ContextCursor
To serialize a Context, we need to get the full information inside a
ContextCursor (and the extra information that only a
ReferenceContextCursor has as well). A side effect of this is that
ApiV2ContextCursor::getStatementArray() can use those getters instead of
receiving the values as three extra parameters.

The getMainArray() implementations are updated to use getters instead of
referring to the fields directly, which is slightly nicer in terms of
code style and also helps adding test coverage for those new methods.
The only getter which isn’t covered in this way (because for a statement
context, statement and snak property ID are the same) gets a test of its
own, so that the whole Context namespace still has full test coverage.

Bug: T185712
Change-Id: Ie600d976676c8bf403ee6dda4ea0055801e1f050