Add icons before statuses on special page
Using the same icons in the gadget and on the special page makes them
easier to learn and improves the understandability of the special page
for people who are already familiar with the icons from the gadget.

The white-space: nowrap rule is necessary to prevent a line break
between the icon and the status text, which is ugly. (For some reason,
using a no-break space between the icon and status text doesn’t prevent
the line break in the table cell.)

Bug: T188728
Change-Id: I3dcb89055b4e838332c13fa1778d29cebf40ab5f