Make CheckResult serializable
Two new classes, CheckResultSerializer and CheckResultDeserializer, are
added, and as for Constraint and ContextCursor, a single test class
covers both of them so we can reuse the test cases.

The DependencyMetadata is serialized by default, but this can be
disabled with a constructor flag, since we don’t need that for the
caching use-case (we store the merged dependency metadata for all
results out-of-band in the cached value, so also storing individual
dependency metadata for each result would just be a waste of space).

Bug: T185714
Change-Id: I81b9db3cdd70c9b94a386c1038e514cd270a9cdf