Replace NullResult’s Context with ContextCursor
NullResult is only used to ensure that the cursors of some contexts
appear in the API output even if they don’t have proper results, so it
makes more sense to restrict it to a context cursor instead of a full
context. This is also required for T189842, which will be resolved by
adding a NullResult with a special context cursor that doesn’t belong to
any full context.

Bug: T189842
Change-Id: Ibb9efd3ef820088c40a1687719f38e191bbe2d87