Ensure that result is populated even for empty entities
We want the API response to be populated at least to the 'claims' level
even for entities that don’t have any statements. We used to partially
ensure this at the CheckConstraints level (but without the 'claims'
part, since this was written for the first API format), but to avoid
“undefined index” errors in CachingResultsBuilder this needs to be
ensured at the CheckingResultsBuilder level already.

DelegatingConstraintChecker::checkAgainstConstraintsOnEntityId() gains a
second callback parameter to populate check results which is called once
per entity (the existing callback is called once per context), and
CheckingResultsBuilder uses this to add a NullResult with an
EntityContextCursor, which populates the 'claims' level of the response.

Bug: T189842
Change-Id: I2d4ef4c2bf5cf6315ae7d7949701553067571613