Make ResultsCache format version configurable
The result format version needs to be bumped from time to time, so it
shouldn’t be hard-coded directly in ResultsCache::makeKey(), but instead
be a constructor parameter. The 'v2.1' value is still hard-coded (now in
ResultsCache::getDefaultInstance()), but now we’ll no longer have to
update various tests when we need to bump the value again.

Bug: T189593
Change-Id: I87bad935906ef1a08706bc77ee66a9aa0ddf264f