Add CachingResultsSource
CachingResultsSource is a wrapper around another ResultsSource which
reads and writes CheckResult serializations to/from cache.

A lot of the code and tests are copied from CachingResultsBuilder, but
with the important difference that the underlying ResultsSource returns
CheckResults which we serialize (or, when reading from cache,
deserialize) instead of dealing with opaque result structure objects.
This also means that we can be a bit more lenient with the status
parameter: if extra statuses were requested, we can still store results
in the cache (filtering out the CheckResults with other statuses), and
if less statuses were requested than we want to cache, we can still
return results from the cache (similarly filtering out CheckResults).

Bug: T189593
Change-Id: I11dc6f78b73afba32c3fb38127bb6c5c7bb9db96