Remove support for plain message CPEs from SparqlHelper
SparqlHelper used to serialize ConstraintParameterExceptions with plain
message strings instead of ViolationMessage objects, which was a bad
idea because the cache is used for all languages, and the plain message
string is language specific. This was fixed in commit e386fbf2a4 (change
I477acc140c), and in commit 1d5be7bc36 (change I49d4170814) some code
was added to SparqlHelper to remove all these “old-style” serializations
from cached arrays. That was over a month ago, and our cache TTL is set
to one day, so any cache map that might have contained “old-style”
serialization should by now either be expired or have been refreshed
(which includes a full removal of “old-style” serializations).
Therefore, I think it’s safe to get rid of this code now.

Bug: T187199
Change-Id: I9e3758d1aa790d2ed83a98f0a7e5e71f2f8de796