Inject MessageLocalizer into message renderers
ConstraintParameterRenderer, ViolationMessageRenderer and
MultilingualTextViolationMessageRenderer now get a MessageLocalizer
injected which they use for creating Message objects, instead of calling
wfMessage() or the Message constructor directly.

For use in tests, a small MessageLocalizer implementation is added,
since no such implementation seems to exist in MediaWiki core.
Currently, it defaults to not setting the language code, but this will
be changed in the future to make the tests more robust.

Bug: T187793
Change-Id: If9f8a1a7f678168d1df95de711a24f5fede5b5e2