Use separators in single, single best, multi value constraints
The three checkers get a ConstraintParameterParser injected and use it
to parse the “separators” parameter, and pass that into
Context::getSnakGroup(). New violation messages for constraints with
separators are added. There are also some other improvements (e. g.
modernizing the existing messages to no longer refer to statements).

A “separators” parameter for a “multi-value” constraint doesn’t really
make too much sense, but we still add it for consistency.

Bug: T193635
Change-Id: I706e603621596de1a054c48d112906c933ce41e0