Don’t try to access null message’s message key
In LoggingHelper, we log long-running constraint checks of any type –
including ones that resulted in a compliance, not a violation. But since
compliance results don’t have a violation message, a
BadMethodCallException would result when trying to get the message key
of the null message. Avoid this and instead just log a null message key
in that case.

Bug: T194140
Change-Id: I05001d78ebbe4737c76c4e737379666fbae17217