Check bounds in IntegerChecker
After some discussion, values like 10±0.5 (integer value, non-integer
bounds) should be rejected as well. This requires some refactorings in
IntegerChecker.

A new message is added, but used only if the value is integer but the
bounds are not. If the value is not integer (e. g. 10.5 or 10.5±0.25),
we use the simpler message that does not mention the bounds.

Bug: T167989
Change-Id: I0cb3341d24435c2c3fb05b70b26f176facadcba1