Add support for property scope message arguments
Property scope arguments are very similar to constraint scope arguments,
but are rendered differently, since property scopes and constraint
scopes are represented by different items. Still, most of the code here
is copied and adapted from the corresponding constraint scope code.

The wbqc-violation-message-property-scope message referenced in the
tests will be added in the upcoming commit that adds support for the
property scope constraint type itself.

Bug: T197473
Change-Id: I80e8a161b0f0655d3f7bf8e028808c3fee4fb51a