Add “property scope” constraint type
This is a more general version of the “used as main value”, “used as
qualifier” and “used as reference” constraint types, and will eventually
replace them (though for some time we support both). It can express all
of those constraint types, but also combinations like “used as main
value or qualifier but not as reference” which previously were only
possible with ugly constructs abusing the constraint scope.

Bug: T197473
Change-Id: I73deee88deb6a82026252209903e6ff682f959b3