Report SPARQL errors as TODO, not VIOLATION
It doesn’t really make sense to show SPARQL errors to users, since they
most likely can’t do anything about it. SparqlHelper already tracks the
errors in statsd (with additional information, e. g. HTTP status code,
if available), so in DelegatingConstraintChecker we don’t need to do any
more logging (the SparqlHelperException contains no extra information).

The TODO status is not the ideal choice for this, but if its
interpretation is widened from “unimplemented” to “could not check due
to implementation fault”, it kinda makes sense. The important thing for
our purposes is that this status is not reported to users by default.

Bug: T199788
Change-Id: I68973a9ec8db587d2059f6ac68a2aa6d613ba7f1