Make index on constraint_guid unique
In commit 8fe3d02510 (change I3201fda882), a new primary key column was
added to the wbqc_constraints table, and an index was added on the
previous primary key, the constraint_guid column, to preserve much of
the behavior of the previous primary key. However, as was pointed out in
T189101#4525809, that index fails to include the UNIQUE constraint which
a primary key implies. The fix for this is simple (change CREATE INDEX
to CREATE UNIQUE INDEX), but to ensure that the updater applies this
change, we have to add the index under a new name, so that we can
distinguish between wikis where the index has been made unique and wikis
that still have the old, non-unique index. (And on that occasion, we
also drop the redundant “index” component from the name, since it is not
usually used in MediaWiki.)

Bug: T180834
Bug: T189101
Change-Id: I642c8924b164b4b8f581b075d64275b75f16e8f8