Replace null SparqlHelper with DummySparqlHelper
Previously, various classes that get a SparqlHelper injected would also
accept null as an argument, which meant that a SPARQL query service was
not available. This causes problems with MediaWiki service registration,
since it is not possible to register null as a service. Instead, a new
class is introduced which is equivalent to a null SparqlHelper.

Bug: T196053
Change-Id: Ie59eb268793c3c3dcd7537ddd7652316d282ce39