Move constraint checkers and DelegatingConstraintChecker to services
Since MediaWikiServices services must be objects, we can’t register a
single array similar to ConstraintReportFactory’s “checker map”;
instead, we register a separate service for each constraint type, and
then another one for the DelegatingConstraintChecker. This is done in a
separate wiring file and with a separate a separate convenience class,
similar to the separation of Wikibase services (see previous commit).

Currently, there is a 1:1-mapping between constraint types and
constraint checker implementations, but this could change in the future.
Several new TODOs in ServiceWiring-ConstraintCheckers.php hint at this,
and that is also the reason why the getters in ConstraintCheckerServices
all declare a generic return type of ConstraintChecker, not the
particular implementation.

The instructions in the README are updated for the new process to
register a checker. The link to the example commit is removed, since it
no longer matches these steps.

Bug: T196053
Change-Id: I4dd44e30bd09b93ec435e1096d9308a4f06b5315