Add more tracking for constraint checks
This adds a method to LoggingHelper to log the total amount of time of
the constraint checks on an entity. The time is tracked via statsd, and
if it exceeds certain configurable limits, also logged at different
levels to the logger interface, analogous to the existing logging for
individual constraint checks. The 'warning' threshold for a full
constraint check is 55 seconds, which is just below the time limit which
is enforced for GET requests in Wikimedia production (60 seconds).

Bug: T204715
Change-Id: Id8a9a4a7a539705893881795b7323cf93fc7a4c0