Respect 429 Http-Responses
Uses Expiry Lock to throttle Sparql requests

The SparqlHelper is now able to identify the throttling (by 429
header-response-code) and will log/count their appearances.

If the response-header is valid, SparqlHelper will block any
SPARQL-query until the time given in the Retry-After header and throw a
TooManySparqlRequestsException.

It also log a warning if the Retry-After header does not exist, is empty
or is invalid and will block any SPARQL-query for a given time (in
seconds), which is provided by
WBQualityConstraintsSparqlThrottlingFallbackDuration.

WBQualityConstraintsSparqlThrottlingFallbackDuration must be a integer,
which is greater than zero.

Also there are new testcases to cover the behavior.

Bug: T204469
Change-Id: I75ea66dbe41bfe14b3d9e26cb01f4497ebe8b1e3