Extract config values in SparqlHelper constructor
Instead of storing the Config object in a property and accessing what we
need from it each time we need something, extract all the values we need
up-front in the constructor. This is probably *slightly* more efficient,
and also allows us to add validation of the config values in a single
place, should we choose to do so. (However, currently no validation is
attempted, making this functionally equivalent to the previous code.)

Change-Id: I3fafb0f16ebee3036c2ad1328eff12250edd79ed