Add maintenance script to import constraint entities
This script imports entities needed to configure constraint checks from
Wikidata if they do not already exist locally (which is detected by
checking whether the locally configured entity ID differs from the
default configuration). Statements and sitelinks are dropped, since
we’re not interested in them. Entities are imported using a new system
user, “WikibaseQualityConstraints importer”; if the import fails because
an entity with a matching fingerprint (item/description) already exists,
that entity is used instead. The resulting entity IDs can be printed as
global variable assignments (can be appended to LocalSettings.php for
simple setups) or in a style suitable for $wgConf setups (though the
values must be merged with existing configuration manually).

The test data was downloaded from the following URL:

> https://www.wikidata.org/wiki/Special:EntityData/Q21503250.json?oldid=802460837

Bug: T209958
Change-Id: I64c68e27c4d9a77f70656bbff8b1a3797b998297