Track number of entities checked in PHP
Before enabling more type checks in PHP, we would like to know how many
entities the PHP type checks currently typically load.

We use a timing metric for this, even though this is not actually a
timing: this automatically gives us maximum values, mean, median, and
various percentiles, which is just what we want. (I found this trick on
the Grafana blog [1].)

[1]: https://grafana.com/blog/2016/03/03/25-graphite-grafana-and-statsd-gotchas/#statsd.timing

Bug: T209504
Change-Id: Idc90322637eb7f0ba4cf6e9108d7238688d2a565