Fix extraction of entity IDs from StorageExceptions
The previous regex to extract an entity ID from a StorageException
message only worked for items, and only for ones in a non-main namespace
(otherwise it would extract [Q123 instead of Q123). Fix both of these
problems, and also extract the code into a separate method so it can be
tested for these issues.

Bug: T209958
Change-Id: I17df1ea889c230652012f12ec9044ebd90dd84d1