Add new constraint status for suggestions
If a constraint statement has a constraint status qualifier “suggestion
constraint” (Q62026391 [1] on Wikidata), then violations of the
constraint are assigned a new 'suggestion' status, which is displayed
using a less severe icon and sorted below other violations.

DelegatingConstraintChecker::downgradeResultStatus is extracted as a
separate method to avoid the cyclomatic complexity of getCheckResultFor
exceeding phpcs’ permitted limit.

Since this is a breaking change to the API output format, the new status
is hidden behind a configuration variable and disabled by default;
unless $wgWBQualityConstraintsEnableSuggestionConstraintStatus is set to
true, the ConstraintParameterParser will pretend not to know about the
'constraint suggestion' item, so the status will not be used.

[1]: https://www.wikidata.org/wiki/Q62026391

Bug: T204439
Change-Id: Ia76615712b2438a65b1bc71bb876df6f1216283e