Track revision ID in constraints update job
At the time that we schedule the job, we know the revision ID that
triggered the update, so let’s make it part of the constraint
parameters, and have the job re-schedule itself for later if the
revision ID it read is older than the one that triggered the job.

This also means that we can afford to read from the replica again rather
than from the master (partially reverting the previous commit, though we
still need an EntityRevisionLookup rather than an EntityLookup),
somewhat reducing query load on the master.

Bug: T223372
Change-Id: I0de23a36a452380173a7a49458aafe63bbe4a7a1