Stop using JsonFileEntityLookup in ConnectionChecker tests
Using JsonFileEntityLookup makes the tests less readable because you
have to consult the JSON files to understand the test code (and many
tests reuse the same JSON files, making it unclear which  parts of the
file are actually relevant to any particular test). It also includes a
hidden dependency on the WikibaseRepo global in order to deserialize the
JSON files into entities, which becomes a problem when migrating the
tests to true unit tests, without a full MediaWiki available.

Instead, use the NewItem and NewStatement helpers to create the test
data within the test’s code. This involves a fair bit of copy+paste to
create similar entities (though often not the same entities, so I’m
skeptical how much of the duplicate-looking code should actually be
extracted), but the JSON files were so large that this still amounts to
a net removal of over 100 lines.

(Using NewItem and NewStatement is still a problem for true unit tests
in the current setup, because they’re part of Wikibase and Wikibase’s
autoloading is special, but hopefully that’s an easier problem to solve
than pulling in the entire WikibaseRepo class.)

Bug: T204704
Bug: T258734
Change-Id: I8d915dd18068aa5ee8ce93f77de959bb203534f4
(cherry picked from commit 9fa2f7ef8521546183da423f58770ef4f03e2b36)