Remove last vestiges of WikibaseQuality
The Travis build still installed the archived WikibaseQuality extension,
which, while no longer necessary, still worked for a time, until the
GitHub mirror was deleted a bit over two weeks ago.

The only other occurrence is in the composer.json description, so we
might as well remove that in the same change.

    $ git grep 'WikibaseQuality[^C\]' | wc -l
    0

Bug: T205065
Change-Id: I12c9605486a9706a7e3a4ad9e6a45890e53a3075