Remove $wgLang mangling from ResultsCacheTest.php
This was originally added in change I5dc2590f30 (commit e63bf4e5e8),
which added the language code to the cache key for constraint check
results, as a stop-gap measure to account for translated constraint
violation messages in the cached results. Since then, we’ve migrated to
caching language-independent results instead, and only localize messages
afterwards, in change Idf4d482818 (commit ff09a1e855). It follows that
we don’t need to adjust the $wgLang in the tests either.

Change-Id: Ia262559b5b98401847e66129284344149a05aaec