Radically reduce ImportConstraintEntitiesTest data
The test data reflected almost the full data of Wikidata’s “type
constraint” item – except for a small reduction in change Ib32aee55b7
(commit 103bc86f7e) – but we need almost none of it. This commit keeps:

* Top-level metadata (item ID, revision ID, etc.)
* English label, which we check in the test
* English description and aliases, which are unused, but don’t cost much
* One sitelink, so we can check that it’s not imported
* One statement, so we can check that it’s not imported

All other labels, descriptions, aliases, descriptions and sitelinks are
removed. On my system, this cuts down the runtime of this test –
previously the slowest test by far of this extension – by a factor of
ten: from ca. 2 seconds to ca. 200 ms.

Change-Id: I7fd1ffd3833cd79fac041df9e324aa874f059f03