Don’t make network requests in DelegatingConstraintCheckerTest
I looked at the WDQS tests with phpspy [1], and it showed that several
tests in DelegatingConstraintCheckerTest actually make network requests
(i. e. they spend some time in curl_exec(), from a flame graph
perspective). This is because the CommonsLinkChecker asks the Wikimedia
Commons API to normalize titles by default, and we don’t mock that
behavior out of it. Let’s do that.

[1]: https://github.com/adsr/phpspy

Change-Id: Icac6b666187ec3046890d2530dd1b95114c4b8b0