Slightly optimize RangeCheckerHelper::__construct()
We only use $this->timeParser to parse timestamps in the standard ISO
format, not arbitrary user input, so we can use a simpler parser
(constructing a full-fledged parser is somewhat expensive). Also, the
TimeValueComparer constructor has an optional parameter to take an
existing TimeValueCalculator instead of creating its own, where we can
pass the one we already have.

Change-Id: Ib1c87754730ca3847f2cf74e4af531829c6bc065