Inject ILoadBalancer into ContraintRepository
Instead of using the wfGetDB() global function (and relying on
MediaWikiIntegrationTestCase manipulating it).

And while we’re at it, let’s update the $db variable names to clarify
what kind of connection we’re talking about ($dbr or $dbw), and use the
ILoadBalancer DB_MASTER and DB_REPLICA constants instead of the globals.

Change-Id: I4b5528e116a5f9f7b299246dfbf8559dfb63daa0