Use ExceptionIgnoringEntityLookup from wikibase/data-model-services
This class was copied to the wikibase/data-model-services package after
being deemed generally useful, so we can use that version and get rid of
our own copy. This does require a version upgrade in composer.json
(3.13.0 is the first version of that package to include the new class),
but Wikibase already requires 3.15.0 and that’s also the version in
mediawiki/vendor, so this shouldn’t be a problem.

Change-Id: I7253a0e6d084bc4a9acd0000ddda377fd6c98d55