Extract public constant in CachingResultsSource
This was not possible at the time the code was introduced (or at least I
believe so), but it’s certainly possible now. Reusing the constant
elsewhere should also prevent bugs like T235110 in the future.

The array_flip() optimization is removed completely. We’re dealing with
very small arrays here, so I don’t think it was/is worth the complexity.

Change-Id: Ic241f651209048661a64e8b7ae5bb3abacbf6445