Inject some services into API modules and special page
The API module and special page registration can specify which services
an API module or special page (class or factory) requires, and pass
those in so they don’t have to be accessed via MediaWikiServices. Do
this for the services that we already get from MediaWikiServices.
(Others are still accessed via global state.)

The $prefix parameter is dropped completely from the API module
constructors and factory methods. I think it was only there by mistake;
usually, the parameter prefix is hard-coded in the parent constructor
call with the API module constructor, not passed around as a parameter
there. (It’s not provided by ObjectFactory.)

Change-Id: Ib44de2c6953e28eb9451db33f7804f4f215b22fd