Don’t cache results with non-local sources
If there are any non-local entity sources, then it’s not safe to use
WikibaseRepo::getLocalRepoWikiPageMetaDataAccessor(), since that might
not support all the entity types we encountered in constraint checking.
Since no non-local version of that service is currently available in
WikibaseRepo (as far as I can see), for now we disable result caching in
that case (after logging a warning, because after all the config asked
for it to be enabled).

Bug: T244726
Change-Id: Ifa001f741ae3fb7b56ca2c13cf62fa130d5f9a87