Handle linking to federated props in constraint violation
ATM the only way to get a link to a federated property is via
'Special:EntityPage'. Because this is a special page it has
namespace == -1, and therefore there is no way to determine its
Talk page. Work around this by using the federated Property
page for the "discuss" link instead

Bug: T239163
Change-Id: Ie905f1800856453ca67732cc1c6d41a59e34c1d8