Handle and track JSON errors in SparqlHelper
If, for some reason, the query service returns a successful HTTP code
but the response isn’t valid JSON, track this in statsd and then throw a
proper SparqlHelperException, instead of crashing with a TypeError when
trying to pass null into the CachedQueryResults constructor.

A PHPCS error about the runQuery() function’s cyclomatic complexity is
temporarily suppressed; I’ll refactor the function in a separate commit,
to make the changes easier to review.

Bug: T260972
Change-Id: I87c57dce62911323452905dd9eb4ff503b2cb01e