Enable SpaceAfterClosure sniff
This one had a lot of violations, but I still think it’s better to
enforce it. The sniff is auto-fixable, and development activity on this
repository is low enough that I think it’s better to enforce more
sniffs, at the cost of temporarily blocking some patches, because
otherwise I fear the code style will slowly drift more and more apart.

Change-Id: I9e0b3e054d0b81b51e82dac06f0d6c747c31ed97