Use lazyPush() for CheckConstraintsJob
If the CheckConstraintsJob can’t be scheduled due to some error, that’s
generally okay and doesn’t require the save to be cancelled, so change
the push() to lazyPush() so that a potential error won’t be thrown in
the hook handler and rollback the transaction.

We keep the lazyPush() for the constraintsTableUpdate job, since that
one seems more important: if we don’t update the constraints table after
an edit to the constraints of a property, many constraint checks will be
incorrect.

Bug: T270522
Change-Id: I12800de16bc1c5608635d416f1a38b7018863113