Use transaction in UpdateConstraintsTableJob
Previously, the update of the constraints table was not wrapped in a
transaction, so (if I’m not mistaken) it was possible for another
request to query the table between the deletion of the old constraints
and the insertion of the new ones, and act as though the property had no
constraints at all. Avoid this by wrapping the update in an atomic
section.

However, to avoid excessive replication lag, we interrupt the
transaction and wait for replication when a hard-coded number of
constraints is exceeded. In this case, a concurrent request might think
the property has fewer constraints than it actually does (but not no
constraints at all, at least). The limit (50) is high enough that for
current Wikidata data, it should never be hit (there are currently no
properties with more than 45 constraint statements, see the query linked
in T226635#5288792).

In combination with the previous commit (change I8ab5a40363), which has
the ImportConstraintStatements script wait for replication after 10 jobs
by default, this means that in the worst case, we will write 500 rows
(10 jobs × 50 rows per job) without waiting for replication. This is
probably on the higher end of the acceptable range, but hopefully not
totally unacceptable yet. (In practice, it is far more common for
properties to have less than 10 constraints, so we would usually wait
for replication after less than 100 rows written.)

Bug: T226635
Change-Id: I4bd2d02ad8a4d85e060b39ed01724521f101c0f7