Don’t inject user language in CheckConstraintParameters
We shouldn’t get the user language at the time the API module is
constructed. Instead, get it from the request context at the time the
module is *executed*, and inject it into a newly created service to
create a ViolationMessageRenderer on demand, using that language.

In the CheckConstraintParameters constructor, add the `string` parameter
type and then remove the doc comment, since it contains no additional
information (and was in fact already outdated).

Bug: T279415
Change-Id: Ia9d6618f03ad03a0330025d68ec04179a623b347