Inject proper context into getApiHelperFactory()
We already have a context source available (the ApiMain), we don’t need
to get the default context from global state.

Note that we need to inject $main->getContext(), not just $main, even
though ApiMain already implements the IContextSource interface required
by getApiHelperFactory() – if we inject $main directly, the constructor
of MediawikiEditFilterHookRunner will complain that it’s being called
with a non-MutableContext.

Change-Id: Ie6d26afb67d45e1abb4ff18cdd6482e2372a0beb