Fix two Phan issues
The StatementList constructor still supports being called with a single
array of Statement objects, but no longer in a form that Phan
recognizes, so let’s use the recommended form and spread the array.

The DataValue interface no longer declares a getHash() method, but all
its implementations still do, so suppress the issue for now. Hopefully a
getHash() declaration will be added back to the interface at some point.

Change-Id: Idcca67bd7dc60b95aab7ff4c6bd93b7f58715d3c