Extract the entity selector hook handler to a factory function
- Write tests for main execution paths of hook handler

This commit extracts the logic involved in the generation of the
entity suggestions list into it's own file. In addition, we employ
dependency injection in order to avoid reliance on any global state
or querying the DOM. Thus, the methods for obtaining the qualifier
context and the main snak's property id are injected as well.

The need to refactor was originally discovered after suggestions
were not working due to a typo, which could have been easily
avoided with some basic unit tests (See I25e8e359c).

In followup commits, this module will be refactored into the
constructor / prototype pattern in order to fall in line with the
pattern exhibited in other modules.

Bug: T283026
Change-Id: I49ac6fe3402297f06a9cdd73ad345dfc764261c6