Handle missing entities in EntitySelectorHookHandlerFactory
It’s unlikely to happen on Wikidata, but in theory the entities
referenced in a constraint could be deleted, in which case wbgetentities
will return an object like { "id": "Q123", "missing": "" }, and we
would get an error when trying to access the labels and descriptions.
Let’s avoid that error.

Change-Id: I5e981d9062c5726cca6c3afbebde885593c25842