Always use supported context types when parsing scope
Previously, we passed the supported (allowed) context types into
parseConstraintScopeParameter() when checking the constraint parameters
(in checkCommonConstraintParameters()), but not when actually checking
the constraints (in handleScope()). This meant that if a constraint
explicitly declared a scope that wasn’t compatible with the constraint
type, then the mismatch would be reported when checking the constraint
parameters, but when checking the constraint, we would just call the
constraint checker on a context type that it doesn’t support (which
usually results in a crash, e.g. when the checker calls
Context::getSnakStatement() and doesn’t expect it to return null). Fix
that by always passing the supported/allowed context types; this also
means that we can make the parameter required in
ConstraintParameterParser.

Bug: T289262
Change-Id: Ibf4826f92a174fc818e0d42dfb2fcfb915503af5