Add ConstraintParameterParser::mapItemId() helper
This simplifies some other methods (especially removing the duplicated
error-construction code), and the mapping arrays are also more flexible
than the previous switch-case, which will be useful for T269724 (where
the allowed values will depend on whether “context types” and “entity
types” are the same “scope” parameter or two separate ones).

In parseRelationParameter(), it’s a bit unfortunate that we only get an
EntityId out of parseEntityIdParameter() and need an ItemId. Previously,
any non-item entity ID would just have been treated like an invalid
(unrecognized) item ID, but I don’t think we should make mapItemId()
accept any entity ID, because it excepts specifically item IDs in the
mapping (otherwise it would need an EntityIdParser to turn the string
mapping keys into entity IDs for the error message). For now, I think
it’s best to copy the parameter-item message from parseItemIdParameter()
(which we can’t use directly because it accepts a PropertyValueSnak, not
an EntityId or a snak serialization). This can surely be cleaned up
later (the various methods in this parser sometimes mix different
“levels” of inputs and outputs in ways that don’t always fully make
sense), but not now.

Bug: T269724
Change-Id: I4f554a38e13229b56d9dedcc177d7516f6fc45d8