Unify terminology around allowed/valid types
We were confusingly using different names for the same thing, the list
of context or entity types that can be used (aren’t “not in scope”) with
a constraint type. The constraint violation message uses the term
“valid”, so let’s use that everywhere in the code as well, even if we
might otherwise slightly prefer “allowed” – the message is much harder
to change than everything else here.

Bug: T269724
Change-Id: I2a1ead4b19c16b96bac5d5e20fe8eeec9a2ee03d